requirements.txt
python-telegram-bot>=22.0
transformers>=4.35.0
torch>=2.0.0
schedule==1.2.0
python-dotenv==1.0.0
aiofiles>=23.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0 